fix-issue-336: update unit test
